TNS-17: parse any top-level 'ns' declaration in file
The ns declaration no longer needs to be the first non-comment form in
the file. This is actually simpler than filtering out 'comment' forms
explicitly.

This also permits read-ns-decl to be called multiple times to search a
file for 'ns' declarations.

'ns' forms still have to be at the top-level of the file.