Port dependency, parse, and track to .cljc
Affects these namespaces:

- clojure.tools.namespace.dependency
- clojure.tools.namespace.parse
- clojure.tools.namespace.track

With minor changes, these files are platform-neutral.

c.t.n.dependency is a general-purpose library which has many
applications outside of tools.namespace.

The uses for c.t.n.parse and c.t.n.track in ClojureScript are less
apparent, but they are easy to port. Perhaps by porting other namespaces
it will be possible in the future to have a version of c.t.n.repl that
works on non-JVM platforms.