Revert "TNS-38: ignore dependency on self"
This reverts commit 6b19f942802f3b35059bb4acfa4791ef1717b8a9.

The commit being reverted here doesn't really solve the problem, just
masks it.

:require-macros is a special case; it crosses the boundary between two
"platforms," Clojure and ClojureScript. As such, it breaks many of the
assumptions on which the current design of tools.namespace is based, in
particular the assumption that there is a one-to-one correspondance
between files and namespaces.

Until some future design in which tools.namespace can understand the
structure of mixed-language projects, I think it makes more sense to
keep each "platform" completely separate, by ignoring
:require-macros (and :use-macros) altogether.